requirements.txt
matplotlib==3.2
numpy==1.19.4
Pillow==8.0.1
-f https://download.pytorch.org/whl/torch_stable.html 
torch==1.8.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html 
torchvision==0.9.0+cpu
ray==1.0.1.post1
pandas==1.2.1
scikit-learn==0.24.1
pickle-mixin==1.0.2
torch-inception-resnet-v2==0.0.33
efficientnet-pytorch==0.7.0

# Flask Framework
Flask==1.1.2
Flask-RESTful==0.3.8
flask-marshmallow==0.14.0
uvicorn
gunicorn